Changes to MBEM, fix tests

+++ src/main/java/com/antigenomics/mageri/core/assemble/DummyMinorCaller.java
@@ -81,7 +81,7 @@ public class DummyMinorCaller extends MinorCaller<DummyMinorCaller> {
     }
 
     @Override
-    public int getGeometricMeanMigSize() {
+    public double getGeometricMeanMigSize() {
         return (int) Math.exp(logMigSize.get() / getTotalMigs());
     }
 }
